+++ src/pixelator/pna/config/panel.py
@@ -21,9 +21,11 @@ import re
 
 import pandas as pd
 import polars as pl
-import ruamel.yaml as yaml
 
-from pixelator.common.config import AntibodyPanelMetadata
+from pixelator.common.config.panel import (
+    AntibodyPanelMetadata,
+    parse_panel_header_metadata,
+)
 from pixelator.common.types import PathType
 from pixelator.common.utils import logger
 
@@ -211,24 +213,7 @@ class PNAAntibodyPanel:
             ValueError: If no metadata header is present in the file.
 
         """
-        yaml_loader = yaml.YAML(typ="safe")
-
-        metadata_lines = []
-        with open(str(file), "r") as f:
-            for line in f:
-                if line.startswith("# "):
-                    metadata_lines.append(line[2:])
-                else:
-                    break
-
-        metadata = "".join(metadata_lines)
-        raw_config = list(yaml_loader.load_all(metadata))
-
-        if len(raw_config) == 0:
-            raise ValueError(f"No header / metadata found in panel file {file}")
-
-        frontmatter = raw_config[0]
-        return AntibodyPanelMetadata.model_validate(frontmatter)
+        return parse_panel_header_metadata(file)
 
     @classmethod
     def _parse_panel(cls, panel_file: Path) -> pd.DataFrame: