+++ tests/mpx/report/__init__.py
@@ -1,5 +1,0 @@
-"""
-Tests for the report.py module
-
-Copyright © 2022 Pixelgen Technologies AB.
-"""